Improved Graphics concept (trait ?)
These changes were made to allow more flexibility with "graphics
object". For instance, user could provide a graphic object that is made
up of a Nz::TextSprite (the actual graphic object that will be used in
interfaces) and a Nz::SimpleTextDrawer (the text renderer). At the
moment, it is how ex::graphics::text is implemented.